fix(generator-pie-component): address PR comments
