+++ packages/tools/generator-pie-component/src/app/types.ts
@@ -1,0 +1,15 @@
+import { Answers } from 'yeoman-generator';
+
+export type TransformedName = {
+  componentName: string;
+  defaultName: string;
+  fileName: string;
+  readmeName: string;
+  percyComponentName: string;
+}
+
+export type Props = {
+  answers: Answers;
+  componentPath: string;
+  storyPath: string;
+} & TransformedName;
