+++ packages/tools/generator-pie-component/src/app/utils.ts
@@ -1,21 +1,39 @@
 import { startCase, snakeCase } from 'lodash';
+import { TransformedName } from './types';
 
 function trim (str: string) {
-    return str.replace(/ /g, '');
+    return str.replace(/\s/g, '');
 }
 
-export function getComponentName (str: string) {
+function getComponentName (str: string) {
     return trim(startCase(str));
 }
 
-export function getComponentFilename (str:string) {
+function getComponentFilename (str:string) {
     return `${trim(str)}`;
 }
 
-export function getReadmeName (str: string) {
+function getReadmeName (str: string) {
     return startCase(str);
 }
 
-export function getPercyComponentName (str: string) {
+function getPercyComponentName (str: string) {
     return snakeCase(str).toUpperCase();
+}
+
+/**
+ *
+ * @param {string} name – String input during prompts for the name of the component
+ * @returns Object – transformed group of names to be used in the generator template files
+ */
+export function transformName (name: string): TransformedName {
+    const normalisedName = name.toLowerCase();
+
+    return {
+        componentName: getComponentName(normalisedName), // e.g IconButton,
+        defaultName: name, // e.g. button
+        fileName: getComponentFilename(normalisedName), // e.g. button.scss
+        readmeName: getReadmeName(normalisedName), // e.g. Icon Button
+        percyComponentName: getPercyComponentName(normalisedName), // e.g. ICON_BUTTON
+    };
 }
